requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.3
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2
matplotlib==3.8.2
Flask==3.0.0
flask-cors==4.0.0